+++ packages/nesd/lib/nes/apu/apu.dart
@@ -261,17 +261,16 @@ class APU {
 
   @pragma('vm:prefer-inline')
   void step() {
-    // triangle and DMC are stepped every CPU cycle
+    // triangle, DMC and the frame counter are stepped every CPU cycle
     triangle.step();
     dmc.step();
+    _frameCounter.step();
 
     if (cycles.isEven) {
       // other channels are stepped every other CPU cycle
       pulse1.step();
       pulse2.step();
       noise.step();
-
-      _frameCounter.step();
     }
 
     if (dmc.startDma) {